feat(tools): add TSDB admin API support behind dangerous flag
Adds support for the TSDB admin APIs defined here:
https://prometheus.io/docs/prometheus/latest/querying/api/#tsdb-admin-apis

These tools are only enabled and registered with the MCP server if the
flag `--dangerous.enable-tsdb-admin-tools` is passed at the command
line. It's long, and it has dangerous in the flag name. Hopefully people
read and understand that carefully before attempting to use these 😅

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>