fix: user agent config on prometheus client
I forgot that setting the client and roundtripper fields on the
prometheus API client are mutually exclusive. This fixes an issue where
the API server fails to initialize. This is a non-fatal error -- we stay
running so that we can continue logging errors (such as these). However,
this does result in tool calls failing due to panic because of the
uninitialized prom API client, so there's probably still room for
improvement in how this is handled. Regardless, this at least fixes
client creation and subsequent tool usage.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>