feat(metrics): add counter for api failures, histogram for api durations
As opposed to the existing metrics for tool call failures/durations
which time the tool, these metrics are specific to the Prometheus API
calls made behind the scenes. This should help make it clear if there
are issues relating to the tool call/llm interaction vs raw api calls to
prometheus.

I realize I could just add another HTTP middleware that times requests
and updates the histogram for me and use that middleware when creating
the Prom API client, but I was already working here for the api error
counter metric and it was modest enough to just instrument here.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>