ref: decouple api creation, move http config loading to main
The API client creation bits were pretty tightly coupled to the HTTP
config loading logic, but there's no need for that part of things to be
done in the API client creation code. So this refactors the HTTP config
file loading into the main package alongside it's flag handling logic,
and includes a small helper function in main to wrap the config loading
logic.

This decoupling should have the added benefit of making it easier to
setup mocks for testing within other packages.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>