fix: add resource middleware to inject api client for resource calls
The switch to injected API clients in contexts in #15 was good for
breaking things up to be less tightly coupled and more testable, but
I only implemented the tool side of the middleware, as tool support has
been my primary focus. Resources need access to the API client as well
in order to work, and since tools and resources use the same underlying
functions to retrieve data from the API, that means resources need a
similar middleware handler to inject the API client.

The upstream mcp-go library doesn't currently have support for resource
middlewares, so this is a basic and custom implementation for our
purpose here. I've got a mock implementation of resource middlewares I'm
testing and considering contributing upstream to the project.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>