feat: add /docs API endpoint to serve static prometheus docs
This is phase 1 of adding documentation to the MCP server.

This commit adds the prometheus/docs repo as a submodule. The docs
subfolder that contains the markdown docs files are then embedded into
the binary, and then the static docs are served under the `/docs` route
on the webserver. This also updates CI in release process to pull
submodules.

Phase 2 is to add a background goroutine to use go-git to periodically
(daily?) check for updates to docs repo and update the docs to serve.

Phase 3 is to make resources to wrap the docs API to serve docs.

Phase 4 is to potentially add new tools to interact with the docs
resources.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>