feat(tools): add tools to list/read docs, add docs tool middleware
The tools call the handler of the appropriate resource. As noted in the
code comments, there's probably a cleaner/better way to have a tool call
a resource using an in-process client or something similar, but the
resources/tools involved here aren't that complicated and this is
functional, so it's good enough for now.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>