ref: accept context when setting up mcp server
Not needed right now, but this will be useful down the line if this is
expanded to include client capabilities as well. Accepting a context in
the mcp server setup func allows us to pass in the root context from
main, so it will inherit cancelations from the root cancel func,
ensuring that any contexts used from here all get cleaned up as
expected.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>