ref: convert tool telemetry from hooks to middlewares, remove sync
This converts the tool telemetry (metrics + logging) from a combination
of before/after tool call hooks into a tool middleware. The existing
tool hooks work, but for operations that require state or some kind of
hand off from before->after, it gets tricky. For instance, timing tool
calls for histograms/logs previousy required a map and mutex to sync up
adding/removing timing entries from the map. This is awkward, requires
extra memory to maintain the map, and forces synchronization where it's
not really needed. Making this a middleware allows for wrapping the tool
call, which is much more ergonomic for things like timings.

This also sets a foundation a resource middleware, which can be added
along with metrics/logging next.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>