feat(mcpserver): add configurable toolsets
This adds the ability to load a specified group of tools at the command
line with the flag `--mcp.tools`.

The available options:

- `--mcp.tools="all"`, this is the default setting and loads all tools.
- `--mcp.tools="core"`, this loads only the minimum required tools to
  interact with prometheus and perform queries.
- `--mcp.tools="X" --mcp.tools="Y" --mcp.tools="Z"`, this loads the core
  tools, as well as the tools specified in the list.

Behind the scenes, this uses a map of mcp tool names and their
corresponding tool implementations. The existing tool implementations
(the variables names, I mean) have been renamed to be more obviously
related to upstream/vanilla prometheus.

Things are structured in a way to allow for room for future provider
specific implementations to tailor the MCP server to systems like
thanos, mimir, cortex, etc -- the idea would be to piggy back the
configurable tool groups and allow provider specific tool mappings,
which could allow for removing tools not supported, substituting
different ones for provider specific implementations, or adding new
tools that are specific to that provider.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>