feat(tools): add docs search
Using the document chunking set up previously, this uses blevesearch v2
to create an in-memory index and index all the chunks.

Indexing is done with the default text analyzer. Search input is also
passed through the same text analyzer and then fuzzines is enabled for
more lenient matchings. The tool accepts a query string and an int param
to limit search results. To return results, the tool adds a default text
snippet that lists the files with matches found, and then a slice of
embedded resources that contain the contents of the files in question so
the LLM can search and read all in one go to find the info it needs.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>