feat: allow customizing tools for select backends, add thanos support
This is a feature I've talked about with several people already, and now
that the groundwork is laid for customizing toolsets loaded to the
server, this adds a further abstraction to allow customizing the tools
loaded by the MCP server for a specific prometheus compatible backend.

This is done by maintaining a mapping of tool name -> {description +
tool implementation}, so it's possible to maintain a mapping of
available tools and their implementations per-backend that we choose to
support.

This commit also adds initial customized support for thanos. Other
backends and tool customizations can be added in future work.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>