ref: add wrapper container to handle per-request api client mods
To consider supporting additional endpoints beyond what the standard
prometheus API client is capable of, I need access to the round tripper
and URL that the client was configured with so I can create requests
with similar authentication setups to the client.

One such use case here is to add a thanos-specific tool for their
endpoint to list store api servers.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>