ref: move prometheus tools to their own file for organization
Now that backend support is a thing and we're preparing to add new
backend-specific tools, organizing tools by their respective backends
will likely be helpful.

This includes the docs tools; while generic in their tooling, they are
specific to prometheus documentation. It's unlikely we'll embed docs for
external backends, as well.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>