feat(tools): add general way to make HTTP request for 3rd party systems
This will be useful for adding new tools for 3rd party prometheus
systems that implement additional endpoints in their API.

This project is still Prometheus-first and all 3rd party systems are
best effort, so ideally, we don't go pulling in api client libs for each
of the various backends. This provides a way to utilize the same round
tripper and url that were used to create the prometheus API client, and
to construct an HTTP request with them against the provided path. This
allows us to ensure that the HTTP client gets created with the same
authorization/credentials/etc as the primary client, so authentication
and access should work the same.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>