feat(tools): add optional response truncation via `--prometheus.truncation-limit=N`
This enables tool response truncation to impose limits on the number of
lines/entries that are returned to the LLM from the API response. It
applies everywhere reasonably possible, such as the query, range_query,
series, label_names, label_names, etc tools. For tools that already
accept a `limit` arg on the tool (such as the target/metric metadata
tools because the underlying API endpoints support a `limit` param), the
truncation limit controls the `limit` arg on the tool.

Default behavior is no truncation. LLMs can override global truncation
behavior on a per-tool-call basis by setting the `truncation_limit` or
`limit` tool arguments as needed. This includes the ability to
temporarily disable truncation by setting the tool arg to `0`. Tool
argument overrides work regardless of whether or not global truncation
has been enabled, meaning that if truncation is left globally disabled,
the LLM can still set a `truncation_limit=100` on a specific tool call
request.

When response truncation has been performed, the MCP server appends a
warning explaining that:
- truncation has occurred
- what the truncation limit is
- suggestions on how the LLM can either optimize the response or re-run
  it with a higher/lower truncation limit or truncation disabled

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>