feat: add `tsdb_blocks` tool
Feature was added in Prometheus v3.6.0, and I added support for it in
prometheus/client_golang#1896. I have a few fixes to that that need to
get into a PR, but the functionality hasn't made it's way into a release
yet, so not a huge rush anyway.

This stages the ground work for the `tsdb_blocks` tool and the follow-on
API client updates for recent API work.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>