feat(tools): support time duration strings in addition to epoch/rfc3339
Explained in more detail in code comment.

TL;DR is that LLMs often like to use duration strings, which often
results in an awkward trial and error loop for the LLM to get the
timestamp into the correct format. Rather than try to fight/train the
LLM, we add support to lean in and play to it's strengths.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>